text_recognition/pytesseract/requirements.txt
pillow
opencv-python
cffi
pandas
defusedxml